\name{statatlas-package}
\alias{statatlas-package}
\alias{statatlas}
\docType{package}
\title{
\packageTitle{statatlas}
}
\description{
\packageDescription{statatlas}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{statatlas}
\packageIndices{statatlas}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{statatlas}

Maintainer: \packageMaintainer{statatlas}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
